Moved tests out of the src folder
